backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
python-dotenv==1.1.1
psycopg[binary]==3.2.9
psycopg-pool==3.2.6
google-genai==1.21.1
pydantic==2.11.7
pydantic-settings==2.10.1
orjson==3.10.18
